require redis client to be passed into ip throttle adapter
